Remove AI script and update various project files
Removed the AI chatbot script prepareAI.sh and provided updates across multiple project files. This includes:

- Addition of a new HTML file for the website root index that provides a performance tracking display.
- Modifications to YoutubeStreamable.cs, deployment.yaml, DJTwojaStara.csproj, InfoController.cs, appsettings.Development.json, PlayerController.cs, README.md, Helpers.cs, MainDbContext.cs, Program.cs, and general module.cs among others.
- Removal of old or no longer relevant lines of codes and addition of new lines for added functionalities and database migrations.
- Addition of a committed dockerfile for application distribution.
- Changes in file permissions as necessary.

The changes were necessary to remove the dependency on the AI chatbot feature which was very demanding and instead focus on a web-based status display and the main functionality of the music bot. The development shift is also in alignment with the orchestration strategy with the inclusion of dockerfile and Kubernetes deployment script. The added implementation for performance snapshots will allow easier tracking of how resources are utilized. Other minor tweaks were done to improve the service overall.