Add track ID, extend EQ settings, enable track removal, and improve UI
This commit introduces several improvements and enhancements to the DJTwojaStara application. Tracks understood by the bot now carry a unique ID property, this change aims to expand future manipulation capabilities of tracks in a session. The update also includes the extension of the EQ settings for a better user experience. Users can now remove tracks from a session and EQ presets have been modulated for a more distinct sound experience. The bot now announces the EQ preset in use for transparency. Further, the user interface has been improved for better readability and also to provide an inviting link to the user's server. This helps in expanding the user's control over the bot's settings.  Significant changes have also been made to the Performance dashboard to make it more descriptive. The bot info command has been removed from GeneralModule.cs as it is no longer required. Lastly, error handling has been added to ensure that the bot provides useful feedback whenever a session is not found.