Add async Task to Preheat method and exception handling in MixerSource
Additional error handling was added to the Read method in MixerSource.cs ensuring any exceptions thrown during the reading of songs will be logged properly. The Preheat method in multiple classes (IStreamable.cs, YoutubeStreamable.cs, and OpusFileStreamable.cs) was updated to be an asynchronous Task in order to enhance the performance and responsiveness of song pre-loading by avoiding blocking. Also, various cleanups and enhancements were made in other classes such as adding a new CreateYTDL method in YTDlpService.cs and adjusting how the Dispose method logs the stack trace in MixerSource.cs.